Add LacayoScene with basic PlayArea
Made my own scene to mess around on to prevent merge conflicts
with the main scene.

Added Util sprites Circle and Square.

Created rough estimate of what the play area would be in terms
of size and of placement in the scene. This is done to make
creating the paddle be much easier since the design is based
around a bended paddle that also moves around the circumference
of the play area.