Add temp paddle sprites
For curved paddles, the sprite is determined based on the half
angle and its movement is limited based on the size of the paddle
which needs to be set in allowedTravelAngle.

Added 5, 10, and 20 degree paddles for temporary use.

Also tested paddles with the EllipseCollider2D and the collision
creation works relatively well for the curved surfaces. Since
this changes from scene to scene, the Paddle prefab was not updated.